Effect over ride works

+++ lib/Effects/Effects/HueSparkle.h
@@ -9,7 +9,7 @@ void hueSparkleEffectLoop(struct EffectDataPacket *effectState)
   const uint8_t hue1 = 128;
   const uint8_t hue2 = 224;
 
-  for (int i = 0; i < LED_CNT; i++)
+  for (uint16_t i = 0; i < LED_CNT; i++)
   {
     const float ledPercent = float(i) / LED_CNT;
     const uint8_t hue = hue1 + (float(hue2 - hue1) * abs((ledPercent / .5) - 1.0));
